refactor: Move RegexGroupProvider to services package
